Renamed MintTokenMotion to DefaultMotion
