+++ src/modules/dashboard/components/ActionsPage/ActionsComponents/index.ts
@@ -1,3 +1,3 @@
 export { default as DefaultAction } from './DefaultAction';
-export { default as RecoverytAction } from './RecoveryAction';
-export { default as MintTokenMotion } from './MintTokenMotion';
+export { default as RecoveryAction } from './RecoveryAction';
+export { default as DefaultMotion } from './DefaultMotion';
